# openapi.yaml
openapi: 3.1.0
info:
  title: Sistema IoT Riego Agrícola
  version: 0.1.0
paths:
...
  /api/v1/gateways:
    get:
      tags:
      - Gateways
      summary: List Gateways
      operationId: list_gateways_api_v1_gateways_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GatewayDetailResponse'
                type: array
                title: Response List Gateways Api V1 Gateways Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - Gateways
      summary: Provision Gateway
      operationId: provision_gateway_api_v1_gateways_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GatewayProvision'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/gateways/{gateway_id}:
    get:
      tags:
      - Gateways
      summary: Get Gateway
      operationId: get_gateway_api_v1_gateways__gateway_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: gateway_id
        in: path
        required: true
        schema:
          type: integer
          title: Gateway Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gateways/{gateway_id}/configuration:
    post:
      tags:
      - Gateways
      summary: Publish Gateway Configuration
      operationId: publish_gateway_configuration_api_v1_gateways__gateway_id__configuration_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: gateway_id
        in: path
        required: true
        schema:
          type: integer
          title: Gateway Id
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishConfigurationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gateways/me/configuration:
    get:
      tags:
      - Gateways
      summary: Poll Gateway Configuration
      operationId: poll_gateway_configuration_api_v1_gateways_me_configuration_get
      parameters:
      - name: X-Config-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
          title: X-Config-Version
      - name: X-Bindings-Revision
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
          title: X-Bindings-Revision
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayConfigurationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gateways/me/binding-candidates:
    post:
      tags:
      - Gateways
      summary: Submit Gateway Binding Candidate
      operationId: submit_gateway_binding_candidate_api_v1_gateways_me_binding_candidates_post
      parameters:
      - name: X-Event-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 128
          - type: 'null'
          title: X-Event-Id
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindingCandidateCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BindingCandidateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gateways/me/binding-candidates/{candidate_id}/confirm:
    post:
      tags:
      - Gateways
      summary: Confirm Gateway Binding Candidate
      operationId: confirm_gateway_binding_candidate_api_v1_gateways_me_binding_candidates__candidate_id__confirm_post
      parameters:
      - name: candidate_id
        in: path
        required: true
        schema:
          type: integer
          title: Candidate Id
      - name: X-Event-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 128
          - type: 'null'
          title: X-Event-Id
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindingConfirmation'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BindingConfirmationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gateways/me/heartbeat:
    post:
      tags:
      - Gateways
      summary: Gateway Heartbeat
      operationId: gateway_heartbeat_api_v1_gateways_me_heartbeat_post
      parameters:
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gateways/me/location:
    post:
      tags:
      - Gateways
      summary: Gateway Location
      description: 'v2 ``location`` operation: the device reports its configured position.'
      operationId: gateway_location_api_v1_gateways_me_location_post
      parameters:
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GatewayLocationCreate'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gateways/{gateway_id}/status:
    get:
      tags:
      - Gateways
      summary: Admin Gateway Status
      operationId: admin_gateway_status_api_v1_gateways__gateway_id__status_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: gateway_id
        in: path
        required: true
        schema:
          type: integer
          title: Gateway Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gateways/{gateway_id}/update-authorizations:
    post:
      tags:
      - Gateways
      summary: Create Update Authorization
      operationId: create_update_authorization_api_v1_gateways__gateway_id__update_authorizations_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: gateway_id
        in: path
        required: true
        schema:
          type: integer
          title: Gateway Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAuthorizationCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateAuthorizationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gateways/me/update-authorization:
    get:
      tags:
      - Gateways
      summary: Get Update Authorization
      operationId: get_update_authorization_api_v1_gateways_me_update_authorization_get
      parameters:
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateAuthorizationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gateways/me/update-confirmations:
    post:
      tags:
      - Gateways
      summary: Confirm Update
      operationId: confirm_update_api_v1_gateways_me_update_confirmations_post
      parameters:
      - name: X-Event-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 128
          - type: 'null'
          title: X-Event-Id
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConfirmationCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateConfirmationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BindingCandidateCreate:
      properties:
        slot_id:
          type: integer
          exclusiveMinimum: 0.0
          title: Slot Id
        logical_node_id:
          type: integer
          exclusiveMinimum: 0.0
          title: Logical Node Id
        irrigation_area_id:
          type: integer
          exclusiveMinimum: 0.0
          title: Irrigation Area Id
        uid:
          type: string
          maxLength: 128
          minLength: 1
          title: Uid
        serial:
          type: string
          maxLength: 100
          minLength: 1
          title: Serial
      additionalProperties: false
      type: object
      required:
      - slot_id
      - logical_node_id
      - irrigation_area_id
      - uid
      - serial
      title: BindingCandidateCreate
    BindingCandidateResponse:
      properties:
        candidate_id:
          type: integer
          title: Candidate Id
        slot_id:
          type: integer
          title: Slot Id
        binding_status:
          type: string
          title: Binding Status
        submitted_at:
          type: string
          format: date-time
          title: Submitted At
      type: object
      required:
      - candidate_id
      - slot_id
      - binding_status
      - submitted_at
      title: BindingCandidateResponse
    BindingConfirmation:
      properties:
        slot_id:
          type: integer
          exclusiveMinimum: 0.0
          title: Slot Id
        technician_confirmed_at:
          type: string
          maxLength: 40
          minLength: 20
          title: Technician Confirmed At
      additionalProperties: false
      type: object
      required:
      - slot_id
      - technician_confirmed_at
      title: BindingConfirmation
    BindingConfirmationResponse:
      properties:
        candidate_id:
          type: integer
          title: Candidate Id
        slot_id:
          type: integer
          title: Slot Id
        binding_status:
          type: string
          title: Binding Status
        confirmed_at:
          type: string
          format: date-time
          title: Confirmed At
      type: object
      required:
      - candidate_id
      - slot_id
      - binding_status
      - confirmed_at
      title: BindingConfirmationResponse
...
    BindingOverlay:
      properties:
        slots:
          items:
            $ref: '#/components/schemas/BindingOverlaySlot'
          type: array
          title: Slots
      type: object
      required:
      - slots
      title: BindingOverlay
...
    ConfigurationSlot:
      properties:
        slot_id:
          type: integer
          title: Slot Id
        logical_node_id:
          type: integer
          title: Logical Node Id
        irrigation_area_id:
          type: integer
          title: Irrigation Area Id
        hardware_profile_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Hardware Profile Code
      type: object
      required:
      - slot_id
      - logical_node_id
      - irrigation_area_id
      - hardware_profile_code
      title: ConfigurationSlot
...
    GatewayConfigurationResponse:
      properties:
        configuration_version:
          type: integer
          title: Configuration Version
        bindings_revision:
          type: integer
          title: Bindings Revision
        property_id:
          type: integer
          title: Property Id
        configuration:
          $ref: '#/components/schemas/GatewayConfigurationSnapshot'
        binding_overlay:
          $ref: '#/components/schemas/BindingOverlay'
        cloud_status:
          type: string
          enum:
          - inactive
          - never_seen
          - recently_seen
          - stale
          - disconnected
          title: Cloud Status
        edge_status:
          type: string
          enum:
          - pending
          - connected
          - delayed
          - disconnected
          title: Edge Status
      type: object
      required:
      - configuration_version
      - bindings_revision
      - property_id
      - configuration
      - binding_overlay
      - cloud_status
      - edge_status
      title: GatewayConfigurationResponse
    GatewayConfigurationSnapshot:
      properties:
        gateway_id:
          type: integer
          title: Gateway Id
        property_id:
          type: integer
          title: Property Id
        slots:
          items:
            $ref: '#/components/schemas/ConfigurationSlot'
          type: array
          title: Slots
      type: object
      required:
      - gateway_id
      - property_id
      - slots
      title: GatewayConfigurationSnapshot
    GatewayDetailResponse:
      properties:
        id:
          type: integer
          title: Id
        property_id:
          type: integer
          title: Property Id
        status:
          type: string
          title: Status
        configuration_version:
          type: integer
          title: Configuration Version
        bindings_revision:
          type: integer
          title: Bindings Revision
        activated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Activated At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
        slots:
          items:
            $ref: '#/components/schemas/GatewaySlotResponse'
          type: array
          title: Slots
      type: object
      required:
      - id
      - property_id
      - status
      - configuration_version
      - bindings_revision
      - slots
      title: GatewayDetailResponse
    GatewayLocation:
      properties:
        latitude:
          type: number
          maximum: 90.0
          minimum: -90.0
          title: Latitude
        longitude:
          type: number
          maximum: 180.0
          minimum: -180.0
          title: Longitude
      additionalProperties: false
      type: object
      required:
      - latitude
      - longitude
      title: GatewayLocation
    GatewayLocationCreate:
      properties:
        location:
          $ref: '#/components/schemas/GatewayLocation'
      additionalProperties: false
      type: object
      required:
      - location
      title: GatewayLocationCreate
      description: 'Body of the v2 ``location`` operation: the device reports where
        it is installed.'
    GatewayProvision:
      properties:
        property_id:
          type: integer
          exclusiveMinimum: 0.0
          title: Property Id
        slots:
          items:
            $ref: '#/components/schemas/GatewaySlotCreate'
          type: array
          title: Slots
      type: object
      required:
      - property_id
      - slots
      title: GatewayProvision
    GatewaySlotCreate:
      properties:
        irrigation_area_id:
          type: integer
          exclusiveMinimum: 0.0
          title: Irrigation Area Id
        hardware_profile_id:
          anyOf:
          - type: integer
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Hardware Profile Id
      type: object
      required:
      - irrigation_area_id
      title: GatewaySlotCreate
    GatewaySlotResponse:
      properties:
        id:
          type: integer
          title: Id
        irrigation_area_id:
          type: integer
          title: Irrigation Area Id
        logical_node_id:
          type: integer
          title: Logical Node Id
        hardware_profile_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Hardware Profile Id
      type: object
      required:
      - id
      - irrigation_area_id
      - logical_node_id
      title: GatewaySlotResponse
    GatewaySlotStatus:
      properties:
        logical_node_id:
          type: integer
          title: Logical Node Id
        irrigation_area_id:
          type: integer
          title: Irrigation Area Id
        binding_status:
          type: string
          title: Binding Status
        latest_reading_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Latest Reading At
        bound_uid:
          anyOf:
          - type: string
          - type: 'null'
          title: Bound Uid
        bound_serial:
          anyOf:
          - type: string
          - type: 'null'
          title: Bound Serial
      type: object
      required:
      - logical_node_id
      - irrigation_area_id
      - binding_status
      title: GatewaySlotStatus
    GatewayStatusResponse:
      properties:
        gateway_id:
          type: integer
          title: Gateway Id
        status:
          type: string
          title: Status
        edge_status:
          type: string
          title: Edge Status
        last_heartbeat_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Heartbeat At
        config_version:
          type: integer
          title: Config Version
        slots:
          items:
            $ref: '#/components/schemas/GatewaySlotStatus'
          type: array
          title: Slots
      type: object
      required:
      - gateway_id
      - status
      - edge_status
      - last_heartbeat_at
      - config_version
      - slots
      title: GatewayStatusResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PublishConfigurationResponse:
      properties:
        gateway_id:
          type: integer
          title: Gateway Id
        property_id:
          type: integer
          title: Property Id
        configuration_version:
          type: integer
          title: Configuration Version
        published_at:
          type: string
          format: date-time
          title: Published At
      type: object
      required:
      - gateway_id
      - property_id
      - configuration_version
      - published_at
      title: PublishConfigurationResponse
...
    UpdateAuthorizationCreate:
      properties:
        image_version:
          type: string
          maxLength: 128
          minLength: 1
          title: Image Version
        image_digest:
          type: string
          maxLength: 128
          minLength: 1
          title: Image Digest
        expires_at:
          type: string
          format: date-time
          title: Expires At
      type: object
      required:
      - image_version
      - image_digest
      - expires_at
      title: UpdateAuthorizationCreate
    UpdateAuthorizationResponse:
      properties:
        authorization_id:
          type: string
          title: Authorization Id
        image_version:
          type: string
          title: Image Version
        image_digest:
          type: string
          title: Image Digest
        expires_at:
          type: string
          format: date-time
          title: Expires At
      type: object
      required:
      - authorization_id
      - image_version
      - image_digest
      - expires_at
      title: UpdateAuthorizationResponse
    UpdateConfirmationCreate:
      properties:
        authorization_id:
          type: string
          maxLength: 64
          minLength: 1
          title: Authorization Id
        image_version:
          type: string
          maxLength: 128
          minLength: 1
          title: Image Version
        image_digest:
          type: string
          maxLength: 128
          minLength: 1
          title: Image Digest
        technician_confirmed_at:
          type: string
          format: date-time
          title: Technician Confirmed At
        result:
          type: string
          pattern: ^confirmed$
          title: Result
      type: object
      required:
      - authorization_id
      - image_version
      - image_digest
      - technician_confirmed_at
      - result
      title: UpdateConfirmationCreate
    UpdateConfirmationResponse:
      properties:
        id:
          type: integer
          title: Id
        authorization_id:
          type: string
          title: Authorization Id
        image_version:
          type: string
          title: Image Version
        image_digest:
          type: string
          title: Image Digest
        technician_confirmed_at:
          type: string
          title: Technician Confirmed At
        result:
          type: string
          title: Result
      type: object
      required:
      - id
      - authorization_id
      - image_version
      - image_digest
      - technician_confirmed_at
      - result
      title: UpdateConfirmationResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/docs-login